# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/google-auth/authUrl/{uid}": {
      "get": {
        "operationId": "GoogleAuthController_getAuthUrl",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "GoogleAuth"
        ]
      }
    },
    "/google-auth/verify": {
      "get": {
        "operationId": "GoogleAuthController_create",
        "parameters": [
          {
            "name": "error",
            "required": false,
            "in": "query",
            "description": "A specfied error(presents and not empty) will cause a validation failure.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Invalid Authorization Code. @ GAuth#1  \nA link to google already exists. @ GAuth#3  \n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDetailDTO"
                },
                "example": {
                  "message": "Invalid Authorization Code.",
                  "code": "GAuth#1"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "tags": [
          "GoogleAuth"
        ]
      }
    },
    "/google-auth/accessToken/{uid}": {
      "get": {
        "operationId": "GoogleAuthController_getAccessToken",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "GoogleAuth"
        ]
      }
    }
  },
  "info": {
    "title": "Sandglass example",
    "description": "The Sandglass API description",
    "version": "0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "APIErrorDetailDTO": {
        "type": "object",
        "properties": {}
      }
    }
  }
}